requirements.txt
amqp==5.3.1
asgiref==3.8.1
bcrypt==4.2.0
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==45.0.6
Django==5.2.5
django-environ==0.11.2
django-grappelli==4.0.1
django-tinymce==4.1.0
djangorestframework==3.15.2
idna==3.10
kombu==5.5.4
numpy==2.1.2
packaging==25.0
pandas==2.2.3
pillow==11.0.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
pytz==2024.2
redis==6.4.0
requests==2.32.4
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
wheel==0.41.3
drf-spectacular~=0.28.0